requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
dataclasses==0.7
fastapi==0.61.2
h11==0.11.0
passlib==1.7.4
pycparser==2.20
pydantic==1.7.2
PyJWT==1.6.1
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.12.2
